// client/src/core/types/form.d.ts
export interface subjectForm {
  color: string
  name: string
  shortForm: string
}

export interface subjectFormErrorTexts {
  color: string
  name: string
  shortForm: string
  valid: boolean
}